+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [ upcoming ]
+
+### Added
+- **Message Search**: Added the ability to search for messages within a folder using IMAP search features. A search input is now available in the message list toolbar.
+
 ## [0.4.2]
 
 ### Added